Add unify CLI log
