+++ package.json
@@ -40,6 +40,7 @@
     "boxen": "~1.2.1",
     "chalk": "~2.1.0",
     "chokidar": "~1.7.0",
+    "clean-stack": "~1.3.0",
     "clear-module": "~2.1.0",
     "debounce": "~1.0.2",
     "get-port": "~3.2.0",
